requirements.txt
Flask==3.1.1
flask-cors==6.0.1
flask-jwt-extended==4.7.1
flask-sqlalchemy==3.1.1
python-dotenv==1.1.1
bcrypt==4.3.0
cryptography==45.0.4
qrcode[pil]==8.2
reportlab==4.2.5
Pillow==11.3.0
requests==2.32.3
boto3==1.35.91
pytest==8.3.4
pytest-cov==6.0.0
bandit==1.8.0

